+++ src/react/templates/app/IndexHTML.js
@@ -1,10 +1,12 @@
 exports.template = () => {
 	return `<!doctype html>
 <html lang="en-GB" prefix="og: http://ogp.me/ns#">
-<head></head>
+<head>
+	<link rel="stylesheet" media="screen" type="text/css" href="/build/index.css" />
+</head>
 <body>
   <div id="application"></div>
-  <script async type="application/javascript" src="./build/index.js"></script>
+  <script async type="application/javascript" src="/build/index.js"></script>
 </body>
 </html>`;
 };